// apps/auth-server/src/types/auth-helpers.ts
export interface EnrichedUser {
  id: string;
  email: string;
  emailVerified: boolean;
  name: string;
  image?: string | null;
  createdAt: Date;
  updatedAt: Date;
  handle?: string | null;
  role?: string;
  twoFactorEnabled?: boolean;
}

export interface EnrichedSessionUser extends EnrichedUser {
  wallet?: string | null;
  poolAddresses?: Record<string, string>;
}